pyproject.toml
[tool.poetry]
name = "hanabi-backend"
version = "0.1.0"
description = "Hanabi game backend"
authors = ["Gianluca Sasanelli <gianluca.sasanelli@gmail.com>"]
package-mode = false  # Add it here, inside the [tool.poetry] section

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
requests = "^2.31.0"
supabase = "^2.15.2"
python-dotenv = "^1.1.0"
alembic = "^1.16.1"
psycopg2-binary = "^2.9.10"
sqlalchemy = "^2.0.41"
sqlmodel = "^0.0.24"
bcrypt = "^4.3.0"
python-jose = "^3.5.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]
ignore = ["UP007", "B008", "B904", "PLW0603", "PLR2004", "N802", "N812", "PLR0911", "PLR0913", "I001"]

[tool.ruff.lint.isort]
lines-between-types = 1
force-single-line = false
combine-as-imports = true
split-on-trailing-comma = true
known-first-party = ["src"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
lines-after-imports = 2

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]